refactor(entities/hostile): Preparacion MVC para posterior implementacion.
