+++ src/roguelike_game/entities/npc/models/monster.py
